changed find func and other little fixes
changed the i to start from 0 and not 1, changed the names of the test functions to include test_ in the begging of the name.

changed the find function:
made a while loop instead of 3 if's inside a for loop, now the function returns a pointer to the index in the list if the value was found , if not it returns NULL , same as if the list is empty.
the function won't print anything anymore inside it, the print statements will be in the main.